Adding _summary module.
